Add scenario_result, result_note and scenario_highlight policies

+++ app/controllers/notes_controller.rb
@@ -1,6 +1,7 @@
 class NotesController < ApplicationController
   def create
     scenario_result = ScenarioResult.find_by_hashid(params[:scenario_result_hashid])
+    authorize scenario_result, :create_note?
     result_note = service.create(scenario_result, note_params)
     json = result_note.prezi.public_json
     render json: json
@@ -8,6 +9,7 @@ class NotesController < ApplicationController
 
   def update
     result_note = ResultNote.find_by_hashid!(params[:id])
+    authorize result_note
     result_note = service.update(result_note, note_params)
     json = result_note.prezi.public_json
     render json: json
@@ -15,6 +17,7 @@ class NotesController < ApplicationController
 
   def destroy
     result_note = ResultNote.find_by_hashid!(params[:id])
+    authorize result_note
     service.destroy(result_note)
     render plain: 'OK'
   end
